+++ src/main/java/janggi/piece/Chariot.java
@@ -1,4 +1,7 @@
-package janggi;
+package janggi.piece;
+
+import janggi.Camp;
+import janggi.Point;
 
 public final class Chariot extends Piece {
