+++ src/main/java/janggi/piece/Piece.java
@@ -1,4 +1,7 @@
-package janggi;
+package janggi.piece;
+
+import janggi.Camp;
+import janggi.Point;
 
 public abstract class Piece {
