+++ src/test/java/janggi/piece/CannonTest.java
@@ -1,7 +1,9 @@
-package janggi;
+package janggi.piece;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import janggi.Camp;
+import janggi.Point;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
